PolyQuant/PolyScalping/requirements.txt
py-clob-client
aiohttp
pydantic
pydantic-settings
python-dotenv
loguru
fastapi
uvicorn[standard]
websockets
web3
